refactor: rename `galleryAttachment.tsx`
Back when the file was made (in
08a6b6a4cff706c1755de9fabcff58b6baafdfb9,
https://github.com/deltachat/deltachat-desktop/pull/1393),
"Gallery" was called "Media".
Nowadays the name "mediaAttachment" doesn't make it clear
that all its stuff belongs to just the Gallery,
one has to figure out which of the two is for the message
and which is for the Gallery.

So let's rename it.
